Move gen to internal
