+++ gen.go
@@ -1,3 +1,3 @@
 package uax29
 
-//go:generate go run gen/main.go
+//go:generate go run internal/gen/main.go
