+++ internal/gen/triegen/compact.go
