+++ internal/gen/triegen/data_test.go
