+++ internal/gen/triegen/print.go
